+++ src/Extension/Table/TableStartParser.php
@@ -63,7 +63,9 @@ final class TableStartParser implements BlockStartParserInterface
     /**
      * @return array<int, string|null>
      *
-     * @psalm-return list<string|null>
+     * @psalm-return array<int, TableCell::ALIGN_*|null>
+     *
+     * @phpstan-return array<int, TableCell::ALIGN_*|null>
      */
     private static function parseSeparator(Cursor $cursor): array
     {
@@ -131,6 +133,10 @@ final class TableStartParser implements BlockStartParserInterface
     }
 
     /**
+     * @psalm-return TableCell::ALIGN_*|null
+     *
+     * @phpstan-return TableCell::ALIGN_*|null
+     *
      * @psalm-pure
      */
     private static function getAlignment(bool $left, bool $right): ?string